Moved files
Moved onboarding.jsx into pages and created a new css folder to hold css files
