+++ pages/onboarding.jsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 //import App from './App.css';
-import TestAPI from '../pages/testAPI';
-import './onboarding.css';
+import TestAPI from './testAPI';
+import '../src/css/onboarding.css';
 
 const Onboarding = () => {
   const [redTeamPlayers, setRedTeamPlayers] = useState(Array.from(Array(20).keys()).map(() => ({ codename: '', idNumber: '' })));
@@ -48,6 +48,7 @@ const Onboarding = () => {
                   value={player.idNumber}
                   onChange={(e) => handleRedTeamChange(index, 'idNumber', e.target.value)}
                   placeholder="ID Number"
+                  
                 />
                 <input
                   type="text"
